+++ tests/integration/test_middleware_pipeline.py
@@ -1,7 +1,7 @@
 import warnings
 from unittest import TestCase
 
-from msgraphcore.core.graph_session import GraphSession
+from msgraphcore.graph_session import GraphSession
 
 from msgraphcore.middleware.authorization_provider import AuthProviderBase
 from msgraphcore.middleware.authorization_handler import AuthorizationHandler
